+++ Demo/Sources/Demos/Demo Screens/HorizontalLayoutViewController.swift
@@ -39,7 +39,7 @@ final class HorizontalLayoutViewController : UIViewController
                     sizing: .fixed(height: 200)
                 )
                 
-                section += Item.list("carousel", sizing: .fixed(height: 200.0)) { horizontal in
+                section += Item.list("carousel-paged", sizing: .fixed(height: 200.0)) { horizontal in
                     
                     horizontal.layout = .paged {
                         $0.direction = .horizontal
@@ -48,47 +48,64 @@ final class HorizontalLayoutViewController : UIViewController
 
                     horizontal += Section("cards") { section in
 
-                        section.layouts.table.columns = .init(count: 2, spacing: 20.0)
+                        section += Item(
+                            CardElement(title: "This is the first card in a paged layout", detail: "Isn't it neat?", color: .white(0.90))
+                        )
 
                         section += Item(
-                            CardElement(title: "This is the first card", detail: "Isn't it neat?", color: .white(0.90)),
-                            sizing: .fixed(height: 300)
+                            CardElement(title: "This is the second card", detail: "Isn't it neat?", color: .white(0.85))
+                        )
+
+                        section += Item(
+                            CardElement(title: "This is the third card", detail: "Isn't it neat?", color: .white(0.80))
+                        )
+
+                        section += Item(
+                            CardElement(title: "This is the fourth card", detail: "Isn't it neat?", color: .white(0.75))
+                        )
+                    }
+                }
+                
+                section += Item.list("carousel-table", sizing: .fixed(height: 200.0)) { horizontal in
+                    
+                    horizontal.layout = .table {
+                        $0.direction = .horizontal
+                        
+                        $0.layout.itemSpacing = 20.0
+                        $0.layout.padding = UIEdgeInsets(top: 0, left: 20, bottom: 0, right: 20)
+                    }
+
+                    horizontal += Section("cards") { section in
+
+                        section += Item(
+                            CardElement(title: "This is the first card in a table layout", detail: "Isn't it neat?", color: .white(0.90)),
+                            sizing: .fixed(width: 200)
                         )
 
                         section += Item(
                             CardElement(title: "This is the second card", detail: "Isn't it neat?", color: .white(0.85)),
-                            sizing: .fixed(height: 300)
+                            sizing: .fixed(width: 200)
                         )
 
                         section += Item(
                             CardElement(title: "This is the third card", detail: "Isn't it neat?", color: .white(0.80)),
-                            sizing: .fixed(height: 300)
+                            sizing: .fixed(width: 200)
                         )
 
                         section += Item(
                             CardElement(title: "This is the fourth card", detail: "Isn't it neat?", color: .white(0.75)),
-                            sizing: .fixed(height: 300)
-                        )
-
-                        section += Item(
-                            CardElement(title: "This is the fifth card", detail: "Isn't it neat?", color: .white(0.70)),
-                            sizing: .fixed(height: 300)
-                        )
-
-                        section += Item(
-                            CardElement(title: "This is the sixth card", detail: "Isn't it neat?", color: .white(0.65)),
-                            sizing: .fixed(height: 300)
+                            sizing: .fixed(width: 200)
                         )
                     }
                 }
                 
                 section += Item(
-                    CardElement(title: "This is the second card", detail: "Isn't it neat?", color: .white(0.95)),
+                    CardElement(title: "This is the fourth card", detail: "Isn't it neat?", color: .white(0.95)),
                     sizing: .fixed(height: 200)
                 )
                 
                 section += Item(
-                    CardElement(title: "This is the third card", detail: "Isn't it neat?", color: .white(0.95)),
+                    CardElement(title: "This is the fifth card", detail: "Isn't it neat?", color: .white(0.95)),
                     sizing: .fixed(height: 200)
                 )
             }