+++ Demo/Sources/Demos/Demo Screens/WidthCustomizationViewController.swift
@@ -58,7 +58,7 @@ final class WidthCustomizationViewController : UIViewController
                 section.layouts.table.width = .custom(.init(
                     padding: HorizontalPadding(uniform: 10.0),
                     width: .atMost(200.0),
-                    alignment: .left
+                    alignment: .leading
                     )
                 )
                 
@@ -77,7 +77,7 @@ final class WidthCustomizationViewController : UIViewController
                         $0.table.width = .custom(.init(
                             padding: HorizontalPadding(uniform: 10.0),
                             width: .atMost(200.0),
-                            alignment: .left
+                            alignment: .leading
                             )
                         )
                     }
@@ -103,7 +103,7 @@ final class WidthCustomizationViewController : UIViewController
                         $0.table.width = .custom(.init(
                             padding: HorizontalPadding(uniform: 10.0),
                             width: .atMost(200.0),
-                            alignment: .right
+                            alignment: .trailing
                             )
                         )
                     }