+++ ListableUI/Sources/Layout/Grid/GridListLayout.swift
@@ -125,7 +125,7 @@ public struct GridAppearance : ListLayoutAppearance
             constraint : WidthConstraint
         ) -> CGFloat
         {
-            let paddedWidth = width - padding.left - padding.right
+            let paddedWidth = width - padding.leading - padding.trailing
             
             return constraint.clamp(paddedWidth)
         }
@@ -273,7 +273,7 @@ final class GridListLayout : ListLayout
         
         let rootWidth = TableAppearance.Layout.width(
             with: viewWidth,
-            padding: HorizontalPadding(left: layout.padding.left, right: layout.padding.right),
+            padding: HorizontalPadding(leading: layout.padding.left, trailing: layout.padding.right),
             constraint: layout.width
         )
         
@@ -291,7 +291,7 @@ final class GridListLayout : ListLayout
         performLayout(for: self.content.header) { header in
             let hasListHeader = self.content.header.isPopulated
             
-            let position = header.layouts.grid.width.position(with: viewSize, defaultWidth: rootWidth)
+            let position = header.layouts.grid.width.position(with: viewSize.width, defaultWidth: rootWidth)
             
             let measureInfo = Sizing.MeasureInfo(
                 sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
@@ -326,7 +326,7 @@ final class GridListLayout : ListLayout
         
         self.content.sections.forEachWithIndex { sectionIndex, isLast, section in
             
-            let sectionPosition = section.layouts.grid.width.position(with: viewSize, defaultWidth: rootWidth)
+            let sectionPosition = section.layouts.grid.width.position(with: viewSize.width, defaultWidth: rootWidth)
             
             //
             // Section Header
@@ -337,7 +337,7 @@ final class GridListLayout : ListLayout
             
             performLayout(for: section.header) { header in
                 let width = header.layouts.grid.width.merge(with: section.layouts.grid.width)
-                let position = width.position(with: viewSize, defaultWidth: sectionPosition.width)
+                let position = width.position(with: viewSize.width, defaultWidth: sectionPosition.width)
                 
                 let measureInfo = Sizing.MeasureInfo(
                     sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
@@ -385,7 +385,7 @@ final class GridListLayout : ListLayout
             
             performLayout(for: section.footer) { footer in
                 let width = footer.layouts.grid.width.merge(with: section.layouts.grid.width)
-                let position = width.position(with: viewSize, defaultWidth: sectionPosition.width)
+                let position = width.position(with: viewSize.width, defaultWidth: sectionPosition.width)
                 
                 let measureInfo = Sizing.MeasureInfo(
                     sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
@@ -426,7 +426,7 @@ final class GridListLayout : ListLayout
         performLayout(for: self.content.footer) { footer in
             let hasFooter = footer.isPopulated
             
-            let position = footer.layouts.grid.width.position(with: viewSize, defaultWidth: rootWidth)
+            let position = footer.layouts.grid.width.position(with: viewSize.width, defaultWidth: rootWidth)
             
             let measureInfo = Sizing.MeasureInfo(
                 sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
@@ -452,7 +452,7 @@ final class GridListLayout : ListLayout
                     
         performLayout(for: self.content.overscrollFooter) { footer in
 
-            let position = footer.layouts.grid.width.position(with: viewSize, defaultWidth: rootWidth)
+            let position = footer.layouts.grid.width.position(with: viewSize.width, defaultWidth: rootWidth)
             
             let measureInfo = Sizing.MeasureInfo(
                 sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),