+++ ListableUI/Sources/Layout/Table/TableListLayout.swift
@@ -104,11 +104,13 @@ public extension LayoutDescription
 /// ```
 public struct TableAppearance : ListLayoutAppearance
 {
-    public var direction: LayoutDirection {
-        .vertical
-    }
+    /// How the layout should flow, either horizontally or vertically.
+    public var direction: LayoutDirection
     
+    /// If sticky section headers should be leveraged in the layout.
     public var stickySectionHeaders : Bool
+    
+    public var contentInsetAdjustmentBehavior : ListLayoutScrollViewProperties.ContentInsetAdjustmentBehavior
     
     /// Default sizing attributes for content in the list.
     public var sizing : Sizing
@@ -122,11 +124,15 @@ public struct TableAppearance : ListLayoutAppearance
         
     /// Creates a new `TableAppearance` object.
     public init(
+        direction : LayoutDirection = .vertical,
         stickySectionHeaders : Bool = true,
+        contentInsetAdjustmentBehavior : ListLayoutScrollViewProperties.ContentInsetAdjustmentBehavior = .automatic,
         sizing : Sizing = Sizing(),
         layout : Layout = Layout()
     ) {
+        self.direction = direction
         self.stickySectionHeaders = stickySectionHeaders
+        self.contentInsetAdjustmentBehavior = contentInsetAdjustmentBehavior
         self.sizing = sizing
         self.layout = layout
     }
@@ -355,7 +361,7 @@ extension TableAppearance
             constraint : WidthConstraint
         ) -> CGFloat
         {
-            let paddedWidth = width - padding.left - padding.right
+            let paddedWidth = width - padding.leading - padding.trailing
             
             return constraint.clamp(paddedWidth)
         }
@@ -415,7 +421,7 @@ final class TableListLayout : ListLayout
     var scrollViewProperties: ListLayoutScrollViewProperties {
         .init(
             isPagingEnabled: false,
-            contentInsetAdjustmentBehavior: .automatic,
+            contentInsetAdjustmentBehavior: self.layoutAppearance.contentInsetAdjustmentBehavior,
             allowsBounceVertical: true,
             allowsBounceHorizontal: true,
             allowsVerticalScrollIndicator: true,
@@ -449,6 +455,52 @@ final class TableListLayout : ListLayout
         
     }
     
+    private func layout(
+        headerFooter : ListLayoutContent.SupplementaryItemInfo,
+        width : CustomWidth,
+        viewWidth : CGFloat,
+        defaultWidth : CGFloat,
+        defaultHeight : CGFloat,
+        contentBottom : CGFloat,
+        after : (ListLayoutContent.SupplementaryItemInfo) -> ()
+    ) {
+        // Determines the position of the content within the list.
+        
+        let position = width.position(
+            with: viewWidth,
+            defaultWidth: defaultWidth
+        )
+
+        // The constraints we'll use to measure the content.
+        
+        let measureInfo = Sizing.MeasureInfo(
+            sizeConstraint: self.direction.size(for: CGSize(width: position.width, height: .greatestFiniteMagnitude)),
+            defaultSize: self.direction.size(for: CGSize(width: 0.0, height: defaultHeight)),
+            direction: self.direction
+        )
+        
+        // Measure the size of the content.
+
+        let size = headerFooter.measurer(measureInfo)
+        
+        // Write the measurement and position out to the header/footer.
+        
+        self.direction.switch(
+            vertical: {
+                headerFooter.x = position.origin
+                headerFooter.size = CGSize(width: position.width, height: size.height)
+                headerFooter.y = contentBottom
+            },
+            horizontal: {
+                headerFooter.y = position.origin
+                headerFooter.size = CGSize(width: size.width, height: position.width)
+                headerFooter.x = contentBottom
+            }
+        )
+        
+        after(headerFooter)
+    }
+    
     func layout(
         delegate : CollectionViewLayoutDelegate?,
         in context : ListLayoutLayoutContext
@@ -456,16 +508,21 @@ final class TableListLayout : ListLayout
         let layout = self.layoutAppearance.layout
         let sizing = self.layoutAppearance.sizing
         
-        let viewSize = context.viewBounds.size
-        let viewWidth = context.viewBounds.width
+        let viewWidth = self.direction.width(for: context.viewBounds.size)
         
-        let rootWidth = CustomWidth.custom(CustomWidth.Custom(
-            padding: HorizontalPadding(left: layout.padding.left, right: layout.padding.right),
+        let rootWidth = CustomWidth.custom(.init(
+            padding: self.direction.switch(
+                vertical: HorizontalPadding(leading: layout.padding.left, trailing: layout.padding.right),
+                horizontal: HorizontalPadding(leading: layout.padding.top, trailing: layout.padding.bottom)
+            ),
             width: layout.width,
             alignment: .center
         ))
 
-        let defaultWidth = rootWidth.position(with: viewSize, defaultWidth: viewSize.width).width
+        let defaultWidth = rootWidth.position(
+            with: viewWidth,
+            defaultWidth: viewWidth
+        ).width
                 
         //
         // Item Positioning
@@ -483,85 +540,69 @@ final class TableListLayout : ListLayout
         // Set Frame Origins
         //
         
-        var lastContentMaxY : CGFloat = 0.0
+        var contentBottom : CGFloat = self.direction.switch(
+            vertical: layout.padding.top,
+            horizontal: layout.padding.left
+        )
         
         //
         // Header
         //
         
-        switch direction {
-        case .vertical:
-            lastContentMaxY += layout.padding.top
-            
-        case .horizontal:
-            lastContentMaxY += layout.padding.left
-        }
-        
-        performLayout(for: self.content.header) { header in
-            let hasListHeader = self.content.header.isPopulated
-            let headerWidth = header.layouts.table.width.merge(with: rootWidth)
-            let position = headerWidth.position(with: viewSize, defaultWidth: defaultWidth)
-            
-            let measureInfo = Sizing.MeasureInfo(
-                sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
-                defaultSize: CGSize(width: 0.0, height: sizing.listHeaderHeight),
-                direction: .vertical
-            )
-            
-            let height = header.measurer(measureInfo).height
-            
-            header.x = position.origin
-            header.size = CGSize(width: position.width, height: height)
-            header.y = lastContentMaxY
-            
-            if hasListHeader {
-                lastContentMaxY = header.defaultFrame.maxY
+        self.layout(
+            headerFooter: self.content.header,
+            width: self.content.header.layouts.table.width.merge(with: rootWidth),
+            viewWidth: viewWidth,
+            defaultWidth: defaultWidth,
+            defaultHeight: sizing.listHeaderHeight,
+            contentBottom: contentBottom,
+            after: { headerFooter in
+                if headerFooter.isPopulated {
+                    contentBottom = self.direction.maxY(for: headerFooter.defaultFrame)
 
-                if self.content.sections.isEmpty == false {
-                    lastContentMaxY += layout.headerToFirstSectionSpacing
+                    if self.content.sections.isEmpty == false {
+                        contentBottom += layout.headerToFirstSectionSpacing
+                    }
                 }
             }
-        }
+        )
         
         //
         // Sections
         //
         
         self.content.sections.forEachWithIndex { sectionIndex, isLast, section in
+            
             let sectionWidth = section.layouts.table.width.merge(with: rootWidth)
-            let sectionPosition = sectionWidth.position(with: viewSize, defaultWidth: defaultWidth)
+            
+            let sectionPosition = sectionWidth.position(
+                with: viewWidth,
+                defaultWidth: defaultWidth
+            )
             
             //
             // Section Header
             //
             
-            let hasSectionHeader = section.header.isPopulated
             let hasSectionFooter = section.footer.isPopulated
             
-            performLayout(for: section.header) { header in
-                let width = header.layouts.table.width.merge(with: sectionWidth)
-                let position = width.position(with: viewSize, defaultWidth: sectionPosition.width)
-                
-                let measureInfo = Sizing.MeasureInfo(
-                    sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
-                    defaultSize: CGSize(width: 0.0, height: sizing.sectionHeaderHeight),
-                    direction: .vertical
-                )
-                
-                let height = header.measurer(measureInfo).height
-                
-                header.x = position.origin
-                header.size = CGSize(width: position.width, height: height)
-                header.y = lastContentMaxY
-                
-                if hasSectionHeader {
-                    lastContentMaxY = section.header.defaultFrame.maxY
-                    
-                    if section.items.isEmpty == false {
-                        lastContentMaxY += layout.sectionHeaderBottomSpacing
+            self.layout(
+                headerFooter: section.header,
+                width: section.header.layouts.table.width.merge(with: sectionWidth),
+                viewWidth: viewWidth,
+                defaultWidth: sectionPosition.width,
+                defaultHeight: sizing.sectionHeaderHeight,
+                contentBottom: contentBottom,
+                after: { header in
+                    if header.isPopulated {
+                        contentBottom = self.direction.maxY(for: header.defaultFrame)
+                        
+                        if section.items.isEmpty == false {
+                            contentBottom += layout.sectionHeaderBottomSpacing
+                        }
                     }
                 }
-            }
+            )
             
             //
             // Section Items
@@ -569,29 +610,45 @@ final class TableListLayout : ListLayout
             
             if section.layouts.table.columns.count == 1 {
                 section.items.forEachWithIndex { itemIndex, isLast, item in
-                    let width = item.layouts.table.width.merge(with: rootWidth)
-                    let itemPosition = width.position(with: viewSize, defaultWidth: sectionPosition.width)
                     
-                    let measureInfo = Sizing.MeasureInfo(
-                        sizeConstraint: CGSize(width: itemPosition.width, height: .greatestFiniteMagnitude),
-                        defaultSize: CGSize(width: 0.0, height: sizing.itemHeight),
-                        direction: .vertical
+                    let width = item.layouts.table.width.merge(with: sectionWidth)
+                    
+                    let itemPosition = width.position(
+                        with: viewWidth,
+                        defaultWidth: sectionPosition.width
                     )
                     
-                    let height = item.measurer(measureInfo).height
+                    let measureInfo = Sizing.MeasureInfo(
+                        sizeConstraint: self.direction.size(for: CGSize(width: itemPosition.width, height: .greatestFiniteMagnitude)),
+                        defaultSize: self.direction.size(for: CGSize(width: 0.0, height: sizing.itemHeight)),
+                        direction: self.direction
+                    )
                     
-                    item.x = itemPosition.origin
-                    item.y = lastContentMaxY
-                    item.size = CGSize(width: itemPosition.width, height: height)
+                    let size = item.measurer(measureInfo)
                     
-                    lastContentMaxY += height
+                    self.direction.switch(
+                        vertical: {
+                            item.x = itemPosition.origin
+                            item.y = contentBottom
+                            item.size = CGSize(width: itemPosition.width, height: size.height)
+                            
+                            contentBottom += size.height
+                        },
+                        horizontal: {
+                            item.x = contentBottom
+                            item.y = itemPosition.origin
+                            item.size = CGSize(width: size.width, height: itemPosition.width)
+                            
+                            contentBottom += size.width
+                        }
+                    )
 
                     if isLast {
                         if hasSectionFooter {
-                            lastContentMaxY += item.layouts.table.itemToSectionFooterSpacing ?? layout.itemToSectionFooterSpacing
+                            contentBottom += item.layouts.table.itemToSectionFooterSpacing ?? layout.itemToSectionFooterSpacing
                         }
                     } else {
-                        lastContentMaxY += item.layouts.table.itemSpacing ?? layout.itemSpacing
+                        contentBottom += item.layouts.table.itemSpacing ?? layout.itemSpacing
                     }
                 }
             } else {
@@ -606,21 +663,32 @@ final class TableListLayout : ListLayout
                     var columnXOrigin = sectionPosition.origin
                     
                     row.forEachWithIndex { columnIndex, isLast, item in
-                        item.x = columnXOrigin
-                        item.y = lastContentMaxY
-                                                
-                        let measureInfo = Sizing.MeasureInfo(
-                            sizeConstraint: CGSize(width: itemWidth, height: .greatestFiniteMagnitude),
-                            defaultSize: CGSize(width: 0.0, height: sizing.itemHeight),
-                            direction: .vertical
+                        
+                        self.direction.switch(
+                            vertical: {
+                                item.x = columnXOrigin
+                                item.y = contentBottom
+                            },
+                            horizontal: {
+                                item.y = columnXOrigin
+                                item.x = contentBottom
+                            }
                         )
                                                 
-                        let height = item.measurer(measureInfo).height
+                        let measureInfo = Sizing.MeasureInfo(
+                            sizeConstraint: self.direction.size(for: CGSize(width: itemWidth, height: .greatestFiniteMagnitude)),
+                            defaultSize: self.direction.size(for: CGSize(width: 0.0, height: sizing.itemHeight)),
+                            direction: self.direction
+                        )
+                                                
+                        let size = item.measurer(measureInfo)
+                        
+                        let height = self.direction.switch(vertical: size.height, horizontal: size.width)
                         
                         let itemSpacing = item.layouts.table.itemSpacing ?? layout.itemSpacing
                         let itemToSectionFooterSpacing = item.layouts.table.itemToSectionFooterSpacing ?? layout.itemToSectionFooterSpacing
                         
-                        item.size = CGSize(width: itemWidth, height: height)
+                        item.size = self.direction.size(for: CGSize(width: itemWidth, height: height))
                         
                         maxHeight = max(height, maxHeight)
                         maxItemSpacing = max(itemSpacing, maxItemSpacing)
@@ -629,14 +697,14 @@ final class TableListLayout : ListLayout
                         columnXOrigin += (itemWidth + section.layouts.table.columns.spacing)
                     }
                     
-                    lastContentMaxY += maxHeight
+                    contentBottom += maxHeight
                     
                     if isLast {
                         if hasSectionFooter {
-                            lastContentMaxY += maxItemToSectionFooterSpacing
+                            contentBottom += maxItemToSectionFooterSpacing
                         }
                     } else {
-                        lastContentMaxY += maxItemSpacing
+                        contentBottom += maxItemSpacing
                     }
                 }
             }
@@ -645,32 +713,25 @@ final class TableListLayout : ListLayout
             // Section Footer
             //
             
-            performLayout(for: section.footer) { footer in
-                let width = footer.layouts.table.width.merge(with: sectionWidth)
-                let position = width.position(with: viewSize, defaultWidth: sectionPosition.width)
-                
-                let measureInfo = Sizing.MeasureInfo(
-                    sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
-                    defaultSize: CGSize(width: 0.0, height: sizing.sectionFooterHeight),
-                    direction: .vertical
-                )
-                
-                let height = footer.measurer(measureInfo).height
-                
-                footer.size = CGSize(width: position.width, height: height)
-                footer.x = position.origin
-                footer.y = lastContentMaxY
-                
-                if hasSectionFooter {
-                    lastContentMaxY = footer.defaultFrame.maxY
+            self.layout(
+                headerFooter: section.footer,
+                width: section.footer.layouts.table.width.merge(with: sectionWidth),
+                viewWidth: viewWidth,
+                defaultWidth: sectionPosition.width,
+                defaultHeight: sizing.sectionFooterHeight,
+                contentBottom: contentBottom,
+                after: { footer in
+                    if footer.isPopulated {
+                        contentBottom = self.direction.maxY(for: footer.defaultFrame)
+                    }
                 }
-            }
+            )
             
             // Add additional padding from config.
             
             if isLast {
                 if self.content.footer.isPopulated {
-                    lastContentMaxY += layout.lastSectionToFooterSpacing
+                    contentBottom += layout.lastSectionToFooterSpacing
                 }
             } else {
                 let additionalSectionSpacing: CGFloat
@@ -682,7 +743,7 @@ final class TableListLayout : ListLayout
                         : layout.interSectionSpacingWithNoFooter
                 }
                 
-                lastContentMaxY += additionalSectionSpacing
+                contentBottom += additionalSectionSpacing
             }
         }
         
@@ -690,58 +751,44 @@ final class TableListLayout : ListLayout
         // Footer
         //
         
-        performLayout(for: self.content.footer) { footer in
-            let hasFooter = footer.isPopulated
-            let footerWidth = footer.layouts.table.width.merge(with: rootWidth)
-            let position = footerWidth.position(with: viewSize, defaultWidth: defaultWidth)
-            
-            let measureInfo = Sizing.MeasureInfo(
-                sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
-                defaultSize: CGSize(width: 0.0, height: sizing.listFooterHeight),
-                direction: .vertical
-            )
-            
-            let height = footer.measurer(measureInfo).height
-            
-            footer.size = CGSize(width: position.width, height: height)
-            footer.x = position.origin
-            footer.y = lastContentMaxY
-            
-            if hasFooter {
-                lastContentMaxY = footer.defaultFrame.maxY
+        self.layout(
+            headerFooter: self.content.footer,
+            width: self.content.footer.layouts.table.width.merge(with: rootWidth),
+            viewWidth: viewWidth,
+            defaultWidth: defaultWidth,
+            defaultHeight: sizing.listFooterHeight,
+            contentBottom: contentBottom,
+            after: { footer in
+                if footer.isPopulated {
+                    contentBottom = self.direction.maxY(for: footer.defaultFrame)
+                }
             }
-        }
+        )
         
-        switch direction {
-        case .vertical: lastContentMaxY += layout.padding.bottom
-        case .horizontal: lastContentMaxY += layout.padding.right
-        }
+        contentBottom += self.direction.switch(
+            vertical: layout.padding.bottom,
+            horizontal: layout.padding.right
+        )
         
         //
         // Overscroll Footer
         //
-                    
-        performLayout(for: self.content.overscrollFooter) { footer in
-            let footerWidth = footer.layouts.table.width.merge(with: rootWidth)
-            let position = footerWidth.position(with: viewSize, defaultWidth: defaultWidth)
-            
-            let measureInfo = Sizing.MeasureInfo(
-                sizeConstraint: CGSize(width: position.width, height: .greatestFiniteMagnitude),
-                defaultSize: CGSize(width: 0.0, height: sizing.overscrollFooterHeight),
-                direction: .vertical
-            )
-            
-            let height = footer.measurer(measureInfo).height
-            
-            footer.x = position.origin
-            footer.size = CGSize(width: position.width, height: height)
-        }
+        
+        self.layout(
+            headerFooter: self.content.overscrollFooter,
+            width: self.content.overscrollFooter.layouts.table.width.merge(with: rootWidth),
+            viewWidth: viewWidth,
+            defaultWidth: defaultWidth,
+            defaultHeight: sizing.overscrollFooterHeight,
+            contentBottom: contentBottom,
+            after: { _ in }
+        )
         
         //
         // Remaining Calculations
         //
         
-        self.content.contentSize = CGSize(width: viewWidth, height: lastContentMaxY)
+        self.content.contentSize = self.direction.size(for: CGSize(width: viewWidth, height: contentBottom))
     }
     
     private func setItemPositions()